requirements.txt
numpy
scipy
mpmath
matplotlib
pyyaml
pandas